Fixed some warnings
Fixed warnings and commented out a line in stencil due to ambiguos operator from json lib.
Still working of the defered renderer

+++ BlueEngine/BlueEngine/BlueCore/Source/Public/BlueCore/Components/ActorComponent.h
@@ -97,6 +97,6 @@ namespace Blue
 		uint64 mInstanceID;
 		Actor* mOwner;
 		uint32 mFlags;
-		uint16 mSize;
+		uint64 mSize;
 	};
 }
